fix(is_template check): swap out call to viml string match function for builtin lua function. (#145)
This is useful for paths that contain tildes that need to be escaped. i.e. if the path to the vault is in an iCloud folder like:
`/Users/person/Library/Mobile Documents/iCloud~md~obsidian/Documents/Notes`

using the viml string match currently throws an error if the user's vault path contains tildes.