Make searching for notes by filename case insensitive
I have a note named "Rob Snyder.md". When I tried to link to it with cmp
like "[[rob " the note wasn't found.

After this change the note is found. Note that `ignore_case` will set the
`--ignore-case` flag in the rg command. However that flag is ignored and
instead `--glob-case-insensitive` is what affects results.